requirements.txt
django==1.9
djangorestframework==3.3.3
mongoengine==0.9
pymongo==2.8
django-rest-framework-mongoengine==3.3.1
django-cors-headers==2.1.0
